@@ -7,14 +7,14 @@ issue_tracker: https://github.com/flutter/flutter/issues?q=is%3Aissue+is%3Aopen+
 version: 0.6.10+3
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=3.0.0"
+  sdk: ^3.3.0
+  flutter: ">=3.19.0"
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_highlight: ^0.7.0
-  markdown: ^5.0.0
+  markdown: ^7.2.1
   meta: ^1.3.0
   path: ^1.8.0
 
