@@ -54,18 +54,17 @@ final MarkdownStyleSheet Function(BuildContext, MarkdownStyleSheetBaseTheme?)
       result = (Platform.isIOS || Platform.isMacOS)
           ? MarkdownStyleSheet.fromCupertinoTheme(CupertinoTheme.of(context))
           : MarkdownStyleSheet.fromTheme(Theme.of(context));
-      break;
     case MarkdownStyleSheetBaseTheme.cupertino:
       result =
           MarkdownStyleSheet.fromCupertinoTheme(CupertinoTheme.of(context));
-      break;
     case MarkdownStyleSheetBaseTheme.material:
+    // ignore: no_default_cases
     default:
       result = MarkdownStyleSheet.fromTheme(Theme.of(context));
   }
 
   return result.copyWith(
-    textScaleFactor: MediaQuery.textScaleFactorOf(context),
+    textScaler: MediaQuery.textScalerOf(context),
   );
 };
 
